requirements.txt
crawl4ai>=0.7.4
beautifulsoup4>=4.12.0
PyYAML>=6.0.0
lxml>=5.0.0
